Calculate and initialize product buy

+++ src/main/java/com/project/JewelryMS/entity/ProductBuy.java
@@ -43,7 +43,7 @@ public class ProductBuy {
     private Integer chi;
 
     @Column(name = "carat")
-    private Integer carat;
+    private Float carat;
 
     @Column(name = "pbStatus")
     private boolean pbStatus;
